Moved craft() from Knife to Tool
